+++ src/entities/session/index.ts
@@ -1,0 +1,1 @@
+export { default as currentSessionStore } from './session';
