+++ src/processes/RequirePermissionCheck/RequirePermissionCheck.tsx
@@ -1,6 +1,6 @@
 import { type FC } from 'react';
 import { Navigate, useLocation } from 'react-router-dom';
-import { currentSessionStore } from '@store/index';
+import { currentSessionStore } from '@entities/session';
 import { observer } from 'mobx-react-lite';
 
 import { type Permissions } from '@entities/session/session';
