+++ src/shared/appError/errorStore.ts
