+++ src/shared/appError/index.ts
@@ -1,0 +1,1 @@
+export { default as errorStore } from './errorStore';
