+++ src/store/index.ts
@@ -1,21 +1,13 @@
-import currentSessionStore from '@entities/session/session';
+// Этот файл в итоге должен исчезнуть вместе со всей папкой store.
+
 import documentsStore from './documents';
 import attributesStore from './attributes';
 import docTypesStore from './documentTypes';
-import errorStore from '@entities/session/errorStore';
 import { injectStores } from '@mobx-devtools/tools';
 
 injectStores({
-  currentUserStore: currentSessionStore,
   documentsStore,
   attributesStore,
-  errorStore,
 });
 
-export {
-  currentSessionStore,
-  documentsStore,
-  attributesStore,
-  docTypesStore,
-  errorStore,
-};
+export { documentsStore, attributesStore, docTypesStore };
